impl Clone for Composable
